fix admin protocol and admin core
