+++ src/protocol/admin/Cargo.toml
@@ -14,6 +14,7 @@ common = { path = "../../common" }
 config = { path = "../../config" }
 logger = { path = "../../logger" }
 metriken = { workspace = true }
+parking_lot = { workspace = true }
 protocol-common = { path = "../../protocol/common" }
 storage-types = { path = "../../storage/types" }
